% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoallthemes.R
\name{autoallthemes}
\alias{autoallthemes}
\title{autoallthemes}
\usage{
autoallthemes(ggplotvisual)
}
\arguments{
\item{ggplotvisual}{Required. ggplot2 visual that you'd like to see in various themes.}
}
\value{
Slowly showcases all available themes for a given visual
}
\description{
autoallthemes
}
\details{
Cycles through all theme packs.
}
\examples{

library(ggplot2)

my_nested_list <- list(id=c(1,2,3,4,5),
weight=c(1,1,1,1,1),
choice1 = c(0,0,1,1,0),
choice2 = c(1,0,0,1,1),
choice3 = c(0,1,0,1,1),
choice4 = c(0,1,0,1,1),
choice5 = c(1,1,1,1,1))

catadat <-  as.data.frame(do.call(cbind, my_nested_list))

p <- ggplot(catadat)

autoallthemes(p)

}
