% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoallthemes.R
\name{autoallthemes}
\alias{autoallthemes}
\title{autoallthemes}
\usage{
autoallthemes(ggplotvisual, HQsave = FALSE)
}
\arguments{
\item{ggplotvisual}{Required. ggplot2 visual that you'd like to see in various themes.}

\item{HQsave}{Defaults to FALSE. Set to TRUE to output high quality raster images of all the plots.}
}
\value{
Slowly showcases all available themes for a given visual
}
\description{
autoallthemes
}
\details{
Cycles through all theme packs.
}
\examples{

lp <- ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3))+
geom_point()+
  guides(color="none")+
  theme(legend.position="none")

autoallthemes(p)

}
