% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautothemes.R
\name{ggautothemes}
\alias{ggautothemes}
\title{ggautothemes}
\usage{
ggautothemes(ggplotvisual, themecollection = "basic1")
}
\arguments{
\item{ggplotvisual}{Required. ggplot2 visual that you'd like to see in various themes.}

\item{themecollection}{Optional. Here you can cycle through the various themes. Your theme options are (in lowercase): basic1 (default), basic2, ggthemes1, ggthemes2, ggthemes3, ggthemes4, ggthemes5, hrbrthemes, ggthemr1 (NOT YET IMPLREMENTED), gthemr2 (NOT YET IMPLEMENTED), ggthemr3 (NOT YET IMPLEMENTED), ggtech, ggdark1, and ggdark2.}
}
\value{
Plot of multiple versions of visual with various themes.
}
\description{
ggautothemes
}
\details{
Once one chooses a plot they like the best, they can use the ______ function to select it, using the title of their preferred plot as an argument.
}
\examples{

library(ggplot2)

p <- ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3)) +
geom_point() +
  guides(color="none") +
  theme(legend.position="none")

ggautothemes(p)

}
